llm-eval-lab/frameworks/ragas-pipeline/requirements.txt
faiss-cpu==1.9.0
google-generativeai==0.8.3
numpy==2.1.3
openai==1.58.1
pytest==8.3.4
python-dotenv==1.0.1
ragas==0.2.10
datasets==3.2.0
streamlit==1.41.1
